+++ src/components/site/build-home-website.tsx
@@ -1,0 +1,5 @@
+import { SiteHome } from "./site-home";
+
+export function BuildHome() {
+  return <SiteHome />;
+}